requirements.txt
astroid==2.5.6
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
factory-boy==3.2.0
Faker==8.2.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
